+++ src/flet_wizards/__init__.py
@@ -38,4 +38,4 @@ __all__ = [
     "WizardTheme",
 ]
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"